Move Section classes into their own package. Remove IConfigurationState
